% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{ExtractQuantitationData}
\alias{ExtractQuantitationData}
\title{Internal method to extract expression data}
\usage{
ExtractQuantitationData(df, scaling.method)
}
\arguments{
\item{df}{data frame containing individual RSEM file output}

\item{scaling.method}{switch containing scaling options}
}
\value{
a scaled vector of counts
}
\description{
Internal method to extract expression data
}
