% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar.R
\name{generate.varid.from.annovar.vff}
\alias{generate.varid.from.annovar.vff}
\title{Helper function to concatenate the unique identifiers for a variant}
\usage{
generate.varid.from.annovar.vff(df)
}
\arguments{
\item{df}{dataframe read from annovar output file}
}
\value{
A vector containing all unique variant identifiers present
     in an annovar variant file.
}
\description{
Helper function to concatenate the unique identifiers for a variant
}
