% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar.R
\name{GroupByGene}
\alias{GroupByGene}
\title{Helper function to group variants by gene.}
\usage{
GroupByGene(df)
}
\arguments{
\item{df}{dataframe read from annovar output file}
}
\value{
A hashmap from gene name -> variant counts
}
\description{
Helper function to group variants by gene.
}
