% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar.R
\name{load.annovar}
\alias{load.annovar}
\title{Load data from annovar-annotated variant calls.}
\usage{
load.annovar(cells, annovar.filenames, germline.filename,
  reduction = "by_gene", somatic.only = TRUE,
  spike.in.regex = "^ERCC-")
}
\arguments{
\item{cells}{character vector containing cell IDs}

\item{annovar.filenames}{character vector containing annovar file paths.}

\item{germline.filename}{file path for annovar file containing germline 
variants. Used as a positive control for single-cell somatic 
variant analysis.}

\item{reduction}{reduction to use from annovar to Seurat assay.
Currently 'by_gene'}

\item{somatic.only}{boolean, default TRUE, produce counts based only
on predicted somatic variants (subtact germline)}

\item{spike.in.regex}{regex selection for 'CHR' vcf field inidcating
synthetic spike-in. Used as a negative control for single-cell
somatic variant analysis. Default '^ERCC-'}
}
\value{
A matrix containing the reduced counts of each ENSG merged from
    all cells supplied in a Seurat-importable format.
}
\description{
\code{load.annovar} loads Annovar data for the provided cell to file mapping.
By default, does not perfom any quality control on variant call files.
\code{load.annovar} should be provided \code{*.variant_function} files
for each cell. If additional exonic functional pruning is required,
\code{load.annovar} will automatically extract the location of the corresponding
\code{*.exonic_variant_function} file locally.
}
\examples{
\dontrun{

load.annovar(c('1', '2'), 
    c('1.anno.variant_function', '2.anno.variant_function')
)

}

}
