% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{load.rsem}
\alias{load.rsem}
\title{Load Data from RSEM transcript quantitation.}
\usage{
load.rsem(cells, rsem.filenames)
}
\arguments{
\item{cells}{character vector containing cell IDs}

\item{rsem.filenames}{character vector containing RSEM file paths.}
}
\value{
A matrix containing the FPKM counts of each ENSG merged from
    all cells supplied in a Seurat-importable format.
}
\description{
\code{load_rsem} loads rsem data for the provided cell to file mapping
}
\examples{

\dontrun{
load_rsem(c('1', '2', '3'), c('1.rsem', '2.rsem', '3.rsem'))
}

}
