% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcr.R
\name{LoadMiXCR}
\alias{LoadMiXCR}
\title{Load data from MiXCR output files}
\usage{
LoadMiXCR(cells, mixcr.filenames)
}
\arguments{
\item{cells}{character vector containing cell IDs}

\item{mixcr.filenames}{character vector containing MiXCR file paths.}
}
\value{
a melange object with all called TCR sequences for the provided cells
}
\description{
Given a particular TCR chain file set (e.g. TCRB / TCRA), load data into
a confidence score matrix where nucleotide sequence of CDR3 is the primary key.
}
