% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{LoadRSEM}
\alias{LoadRSEM}
\title{Load Data from RSEM transcript quantitation.}
\usage{
LoadRSEM(cells, rsem.filenames, quantitation.method = "TPM",
  min.quant.value = 0)
}
\arguments{
\item{cells}{character vector containing cell IDs}

\item{rsem.filenames}{character vector containing RSEM file paths.}

\item{quantitation.method}{String, default 'TPM', returns expected
counts from RSEM files. 'count' also supported.}

\item{min.quant.value}{numeric, containing the minimum quantitation
a gene must have to be included in the output \code{@data} matrix}
}
\value{
A matrix containing the TPM counts of each ENSG merged from
    all cells supplied in a Seurat-importable format.
}
\description{
\code{LoadRSEM} loads rsem data for the provided cell to file mapping
}
\examples{

\dontrun{
LoadRSEM(c('1', '2', '3'), c('1.rsem', '2.rsem', '3.rsem'))
}

}
