% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{LoadVCF}
\alias{LoadVCF}
\title{Load data from VCF output files}
\usage{
LoadVCF(cells, vcf.filenames, consolidation.factor = 10000)
}
\arguments{
\item{cells}{character vector containing cell IDs}

\item{vcf.filenames}{character vector containing VCF file paths.}

\item{consolidation.factor}{the factor to use for dividing}
}
\value{
a melange object with consolidated variants by consolidation.factor
        rows will be named with the form: _<chr_name>:<position>
}
\description{
Combine variants by position only, to use as input for describing positions of
mutational frequency.
}
