% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Melange}
\alias{Melange}
\title{The Melange Class}
\usage{
Melange(data.matrix)

Melange(data.matrix)
}
\arguments{
\item{data.matrix}{Matrix with data for the specific data type interrogated}
}
\description{
The Melange object stores the cell-specific information aggregated from whichever
tool the user has chosen to import from. The number of rows (features) as well
as the datatype of the core data matrix depends on the particular input file
type,
}
\section{Slots}{

\describe{
\item{\code{data}}{Raw counts data for the loaded data source. Can be any 
matrix.}

\item{\code{source}}{Unambiguous name for the source data in the Melange object.}

\item{\code{meta.features}}{Feature-level metadata}

\item{\code{meta.data}}{Cell-level metadata}

\item{\code{misc}}{Utility slot for storing additional data}
}}

