% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrapper.R
\name{SeuratAssayFromMelange}
\alias{SeuratAssayFromMelange}
\title{Convert data from Melange to Seurat Assay}
\usage{
SeuratAssayFromMelange(object)
}
\arguments{
\item{object}{melange object to be converted into Seurat Assay form for
easier downstream processing and data integration.}
}
\value{
A Seurat Assay object containing the data collected in the Melange
    object.
}
\description{
Melange operations return a melange data object for interoperability
with different analysis pipelines, but many users may prefer to do
single-cell analysis within Seurat. As such, the \code{SeuratAssayFromMelange}
function provides an easy way to convert between the two formats.
Mostly, this is done by re-naming and re-organizing some data elements.
}
\examples{
\dontrun{

SeuratAssayFromMelange(melange.object)

}

}
