% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{LoadRSEM}
\alias{LoadRSEM}
\title{Load Data from RSEM transcript quantitation.}
\usage{
LoadRSEM(cells, rsem.filenames, quantitation.method = "count")
}
\arguments{
\item{cells}{character vector containing cell IDs}

\item{rsem.filenames}{character vector containing RSEM file paths.}

\item{quantitation.method}{String, default 'count', returns expected
counts from RSEM files. 'TPM' also supported.}
}
\value{
A matrix containing the FPKM counts of each ENSG merged from
    all cells supplied in a Seurat-importable format.
}
\description{
\code{LoadRSEM} loads rsem data for the provided cell to file mapping
}
\examples{

\dontrun{
LoadRSEM(c('1', '2', '3'), c('1.rsem', '2.rsem', '3.rsem'))
}

}
